refactor: split control plane snapshot mapping
